Updates regarding file handling
1) Correctly handled optional files zip generation
2) Correctly handled the generated file to users download folder if he wishes to.
3) Moved the packaging operation and IO handling to a Task.
4) Overall tool fixes and improvements.

+++ Assets/_Scripts/_Core/SubmissionHandler.cs
@@ -6,13 +6,16 @@ namespace BetaMax.Core
     using BetaMax.UI;
     using BetaMax.Core.IO;
     using System.IO;
+    using System.Threading.Tasks;
 
     public class SubmissionHandler : MonoBehaviour
     {
         static SubmissionHandler _s;
         public static SubmissionHandler S;
 
+        public string JSON_FILE_NAME = "testerInfo.json";
         public string TEMP_FOLDER_NAME = "SubTemp";
+        public string USER_OPT_FOLDER = "UserOptFiles";
         public string MAIN_ZIP_NAME = "mainZip.zip";
         public string OPT_ZIP_NAME = "optZip.zip";
 
@@ -76,21 +79,53 @@ namespace BetaMax.Core
         #region SUBMISSION_SEQUENCE
         void PackagingSequence()
         {
-            if (ManageAuxProccess())
-            { Debug.Log("Aux process completed"); }
+            string userJsonFolder = Path.Combine(Application.dataPath, TEMP_FOLDER_NAME, JSON_FILE_NAME);
 
-            //collect optional data
-            string optFilesPath = configPanel.OptionalFilesPath;
-            if (!String.IsNullOrEmpty(optFilesPath) | !String.IsNullOrWhiteSpace(optFilesPath))
+            TesterInfo tempInfo = InfoSerialization.DeserializeJsonToObj(userJsonFolder);
+            string optFilesPath = tempInfo.optionalsPath;
+            string downFilesPath = tempInfo.downloadPath;
+            bool canDownload = tempInfo.onsSubmitDownloadValue;
+
+            Task.Run(() =>
             {
-                string finalOptZip = Path.Combine(Application.dataPath, OPT_ZIP_NAME);
-                if (File.Exists(finalOptZip))
-                { File.Delete(finalOptZip); }
+                if (ManageAuxProccess())
+                { Debug.Log("Aux process completed"); }
 
-                CompressionHandler.CompressToZip(optFilesPath, finalOptZip);
-            }
+                if (!String.IsNullOrEmpty(optFilesPath) && !String.IsNullOrWhiteSpace(optFilesPath))
+                {
+                    string userTempFolder = Path.Combine(Application.dataPath, TEMP_FOLDER_NAME, USER_OPT_FOLDER);
+                    string finalOptZip = Path.Combine(Application.dataPath, TEMP_FOLDER_NAME, USER_OPT_FOLDER, OPT_ZIP_NAME);
 
-            //TODO: Package the whole tempfolder
+                    if (!Directory.Exists(userTempFolder))
+                    { Directory.CreateDirectory(userTempFolder); }
+                    if (File.Exists(finalOptZip))
+                    { File.Delete(finalOptZip); }
+
+                    Debug.Log("Started packaging");
+                    CompressionHandler.CompressToZip(optFilesPath, finalOptZip);
+
+                    while (Directory.GetFiles(userTempFolder + Path.DirectorySeparatorChar).Length <= 0)
+                    { }
+
+                    Debug.Log("Finished file generation");
+                }
+
+            }).ContinueWith((t) =>
+            {
+                if (canDownload)
+                {
+                    string userTempFolder = Path.Combine(Application.dataPath, TEMP_FOLDER_NAME, USER_OPT_FOLDER, OPT_ZIP_NAME);
+                    string destPath = Path.Combine(downFilesPath, OPT_ZIP_NAME);
+
+                    if (Directory.Exists(downFilesPath))
+                    {
+                        Debug.Log("downloads exists");
+
+                        File.Copy(userTempFolder, destPath);
+                    }
+                }
+            });
+
         }
 
         bool ManageAuxProccess()
